Refactor and reorganize project structure
Deleted endpoint definitions and refactored directory structure for improved modularity. Renamed and relocated modules from `pan_scm_sdk` to `scm`, enhancing the organization and maintainability of the codebase. Updated import paths accordingly.
