--- pan_scm_sdk/__init__.py
@@ -1,1 +1,0 @@
-# pan_scm_sdk/__init__.py
