--- pan_scm_sdk/auth/__init__.py
@@ -1,1 +1,0 @@
-# pan_scm_sdk/auth/__init__.py
