--- pan_scm_sdk/endpoints/address_groups.py
@@ -1,9 +1,0 @@
-# pan_scm_sdk/endpoints/addresses.py
-
-ADDRESS_GROUPS_ENDPOINTS = {
-    "list_address_groups": "/config/objects/v1/address-groups",
-    "create_address_groups": "/config/objects/v1/address-groups",
-    "get_address_group": "/config/objects/v1/address-groups/{id}",
-    "update_address_group": "/config/objects/v1/address-groups/{id}",
-    "delete_address_group": "/config/objects/v1/address-groups/{id}",
-}
