--- pan_scm_sdk/endpoints/addresses.py
@@ -1,9 +1,0 @@
-# pan_scm_sdk/endpoints/addresses.py
-
-ADDRESSES_ENDPOINTS = {
-    "list_addresses": "/config/objects/v1/addresses",
-    "create_address": "/config/objects/v1/addresses",
-    "get_address": "/config/objects/v1/addresses/{id}",
-    "update_address": "/config/objects/v1/addresses/{id}",
-    "delete_address": "/config/objects/v1/addresses/{id}",
-}
