--- pan_scm_sdk/endpoints/applications.py
@@ -1,9 +1,0 @@
-# pan_scm_sdk/endpoints/applications.py
-
-APPLICATIONS_ENDPOINTS = {
-    "list_applications": "/config/objects/v1/applications",
-    "create_applications": "/config/objects/v1/applications",
-    "get_application": "/config/objects/v1/applications/{id}",
-    "update_application": "/config/objects/v1/addresses/{id}",
-    "delete_application": "/config/objects/v1/addresses/{id}",
-}
