--- pan_scm_sdk/models/__init__.py
@@ -1,1 +1,0 @@
-# pan_scm_sdk/models/__init__.py
