--- pan_scm_sdk/utils/__init__.py
@@ -1,1 +1,0 @@
-# pan_scm_sdk/utils/__init__.py
